removing sph-main, only keeping mpi version
