+++ examples/CMakeLists.txt
@@ -21,5 +21,4 @@ add_custom_target(
 
 add_subdirectory(md-flexible)
 add_subdirectory(sph)
-add_subdirectory(sph-mpi)
 add_subdirectory(sphDiagramGeneration)
