--- examples/sph-mpi/README
@@ -1,2 +1,0 @@
-strongly aligned with sph example from FDPS library
-https://github.com/FDPS/FDPS/tree/master/sample/c%2B%2B/sph
